Wrote helper functions for slot and header regions.
